DEV-302: improve robustness of RDS tool

+++ acceptance-amazon-rds-sync/execute-migration.sh
@@ -9,30 +9,34 @@ DIR=`dirname $0`
 
 source $DIR/config
 
-# Reset 'migration' folder to origin/master
-cd $DATABASE_MIGRATION_SOURCE_DIR ; git stash clear && git stash save ; git fetch -q origin && git checkout -q master
+# skip old-style migrations if source and target directories are not present
+if [ -d "$DATABASE_MIGRATION_SOURCE_DIR" -a -d "$DATABASE_MIGRATION_TARGET_DIR" ]; then
+  # Reset 'migration' folder to origin/master
+  cd $DATABASE_MIGRATION_SOURCE_DIR ; git stash clear && git stash save ; git fetch -q origin && git checkout -q master
 
-# Split file names on newline
-IFS=$'\n'
+  # Split file names on newline
+  IFS=$'\n'
 
-for sql_file_path in `ls -1 $DATABASE_MIGRATION_TARGET_DIR/*.sql`; do
+  for sql_file_path in `ls -1 $DATABASE_MIGRATION_TARGET_DIR/*.sql`; do
 
-	# Get SQL migration file name
-    sql_file=$(basename $sql_file_path)
+  	# Get SQL migration file name
+      sql_file=$(basename $sql_file_path)
 
-	if [ ! -f "$DATABASE_MIGRATION_SOURCE_DIR/$sql_file" ]; then
-		# Execute new MySQL file
-		mysql -h $DATABASE_HOST $DATABASE_NAME < $sql_file_path 1>/dev/null
+  	if [ ! -f "$DATABASE_MIGRATION_SOURCE_DIR/$sql_file" ]; then
+  		# Execute new MySQL file
+  		mysql -h $DATABASE_HOST $DATABASE_NAME < $sql_file_path 1>/dev/null
 
-		# Check for MySQL migration result
-		if [ $? -gt 0 ]; then
-			echo "ERROR - Failed to execute $sql_file!" >&2
-		fi
+  		# Check for MySQL migration result
+  		if [ $? -gt 0 ]; then
+  			echo "ERROR - Failed to execute $sql_file!" >&2
+  		fi
     fi
-done
+  done
 
-# Reset Git working folder to previous branch
-cd $DATABASE_MIGRATION_SOURCE_DIR && git checkout -q - && git stash pop
+
+  # Reset Git working folder to previous branch
+  cd $DATABASE_MIGRATION_SOURCE_DIR && git checkout -q - && git stash pop
+fi
 
 # run migration scripts with scripts/migrate if available
 # the two vars should be set up in the config file
